data_ingestion/requirements.txt
azure-eventhub==5.14.0
pandas==2.1.0
numpy==1.26.0
faker==23.3.0
